Convert Firebase Function to Express app
