+++ functions/package.json
@@ -17,6 +17,7 @@
     "@google-cloud/functions-framework": "^3.3.0",
     "@google-cloud/secret-manager": "^5.0.1",
     "axios": "^1.5.1",
+    "express": "^4.18.2",
     "firebase-admin": "^11.8.0",
     "firebase-functions": "^4.3.1"
   },
