+++ functions/utils/constants.js
@@ -1,0 +1,11 @@
+const {
+  auth0ClientId,
+  auth0ClientSecret,
+  auth0Domain
+} = process.env;
+
+module.exports = {
+  auth0ClientId,
+  auth0ClientSecret,
+  auth0Domain
+};
